Fix all 182 clang-tidy warnings and bump version to 1.0.3-dev
- Remove ~80 unused includes across 30+ files
- Change Ptr<T> params to const Ptr<T>& for performance (15+ files)
- Fix narrowing conversions with static_cast (~10 locations)
- Remove 11 unused using declarations
- Fix widening multiplication and return value warnings
- Replace std::endl with '\n' in FusionLogger
- Suppress array index warnings for intentional TriArc design
- Update VERSION to 1.0.3-dev for git hash in version string

+++ src/adapters/FusionWorkspaceProfile.cpp
@@ -132,7 +132,7 @@ bool FusionWorkspace::extractProfileVertices(const std::string& entityId,
     if (reversed) {
       // Add points in reverse order, skip first point (which is last in
       // reverse)
-      for (int j = strokePoints.size() - 1; j >= 1; --j) {
+      for (int j = static_cast<int>(strokePoints.size()) - 1; j >= 1; --j) {
         if (strokePoints[j]) {
           double x = strokePoints[j]->x();
           double y = strokePoints[j]->y();
